Remove unit test that was system-dependent
Fixes issue #50. The test in question assumed
that a JKS key store would return its entries
in a deterministic order when an alias was not
specified, but this may not be true. Removing
the test entirely since there is already a test
that verifies that the functionality is correct
when an alias is explicitly specified.