Add ext.feign-okhttp3 package
ext.feign-okhttp3 is a copy of feign.okhttp.OkHttpClient in the feign.okhttp3 package.
This constructions allows clients compiled against Feign <= 8.16 (with OkHttp2) to be used
alongside clients compiled against Feign 8.17 (with OkHttp3).

This should be considered a temporary work-around and be removed as soon as Feign
sorts out their back-compat story.