Allow charset for text/plain Content-Type (#216)
* Allow charset for text/plain Content-Type

For many services, the Content-Type includes the charset, i.e.
`Content-Type: text/plain; charset=utf-8`. According to the
Content-Type spec, this is perfectly acceptable. The TextDelegateDecoder
was checking for exact string equality, which resulted in attempting
to use the Json delegate Decoder, as opposed to the String decoder in
this situation.

This commit:

1: Added a failing test to demonstrate the issue,
2: Fixes the failing test, but changing `.equals()` to `.contains()`

* Check Content-Type with startsWith, not contains