Use double checked locking instead of a broken atomicboolean
There's currently a race in the Http2Agent where one can call it
from multiple threads, and then one of the threads exits before
the boolean is installed.

Would ordinarily use Suppliers.memoize, but it seems we don't want
a guava dependency here; and since this wasn't originally synchronized,
I figured this is something where we actively want to avoid
synchronization, so I just made it do volatile reads.