Retrofit can return CompletableFutures
Remote calls are currently indistinguishable from local calls, which
makes them hard to use.

Additionally, Retrofit currently sucks to use; the Call api is hard
to use and throws IOExceptions.

This lets it return CompletableFutures, which is better than Feign
and

All of the AsyncCallTracker crap is just there to maintain backcompat;
the current method used to throw a RemoteException is wholly
incompatible with anything async.