IOException not IllegalStateException
OkHttp interceptors are not allowed to throw IllegalStateExceptions.

The contract of an OkHttp call is...

```
   * @throws IOException if the request could not be executed due to cancellation, a connectivity
   * problem or timeout. Because networks can fail during an exchange, it is possible that the
   * remote server accepted the request before the failure.
```

which throwing an IllegalStateException in an interceptor breaks. This breaks assumptions
in OkHttp, and notably will cause callbacks to never be triggered.

I didn't notice this when I added the CompletableFuture retrofit
support (only noticed the other interceptor).

Fixes #504.