Better behaviour in the presence of backing off
I assert that:

If a client sends requests at a rate of 40 per second, but the server
demands only 20 per second by sending back 429s, the default behaviour
should be that the server should smooth the client's load down to 20
per second.

Right now this does not work, because we use exponential backoff with
a maximum number of retries, and each call is independent, so each
individual request is as likely to succeed as any other.

What this means is that given sustained attempted load of 40 per second
and a limit of 20 requests per second, eventually some of the calls
_will_ fail.

You can see this most easily right now with an endpoint that takes 1ms
to return with a rate limit of 1 request per second, which will fail
within a few seconds.

This PR attempts to fix the issue. It uses Netflix's concurrency-limits
library to apply a tcp-like strategy to request concurrency; basically
provided a single request per second won't be limited into oblivion,
you can make progress.

I also include a test that will fail immediately unless this concurrency
limiting code exists.

At the moment, the scope is per-endpoint - I don't know how to do
better. We could later on define 'limiting domains' which get somehow
annotated by the service author according to rate limit groupings.

Note that we'd have to modify our internal retrofit codegen to properly
add the path template header.

There is scope to improve this later - particularly to do this on the
server too, and prioritise interactive over batch traffic.

The main downside of this code is that you have to make sure you always
clean up - otherwise you have a resource leak. In order to avoid this,
there is code which cleans up in the case of the resource leak, logging
at the same time. This should avoid the IOException deadlock pain
we've seen in the past.