[fix] Empty bodies are appropriately accounted for.
There exists a bug in the old version of Feign we use. If the body
is empty, Feign does not close it. In practice, this means that
zero-length responses are not closed,and so concurrency permits
are not returned.

We still need to handle the null case, but it should be an ignore,
not a success.

Additionally, we should probably consider having some better ETE style
tests for Conjure-java-runtime. This bug was trivial to reproduce
internally, but the tooling on OSS makes it quite hard.