[improvement] Lower the max requests per host (#867)
## Before this PR
A single slow host could block requests to other hosts. This manifested as a result of rate limiting, where a requests would get scheduled to run, then block client side due to exceeding the server's rate limit. Because the OkHttp client is shared across hosts and services, requests to the single host that was rate limiting a client took up all 256 slots in the dispatcher which meant requests to other hosts that were not rate limited were not scheduled.

## After this PR
A single host is only able to take up 1/4th of the dispatcher slots which means whether it's slow or rate limited, requests to other hosts will continue to be executed.