Only select a URL once per request attempt (#960)
Fixes #942

## Before this PR
When using the round robin node selection strategy, retried requests would advance the current URL twice - once in the response handler and again in the URL selector interceptor.

## After this PR
Retried requests will only advance the current URL once, as expected. This PR also simplifies the URL selection by removing the URL selector interceptor and choosing the correct URL when a new request is first attempted.

The URL selector interceptor seems broken in a number of other ways. For example, the current URL can change between when a redirect URL is selected on failure and when the retried request passes through the interceptor, causing a different URL to be used than the one selected in the failure handler.