Fix a couple of concurrency limiting bugs
I was in the area, and stopped by to fix a few bugs.

The first is that the tracing is not set correctly on retries. This
means that RPCs are associated with incorrect traces. We've seen
symptoms of this internally. Use the DeferredTracer to fix this.

The second is that we use a SettableFuture and setFuture on it within
the limiter listener request tag. If the future is already done (e.g.
this is a retry) then this will already be done, and we will not react
properly.

Interestingly this actually has no effect because the Netflix library
does not validate that you didn't call it twice and behaves as though
it was a different listener. But this could change at any time, and
then we'd be in a tough spot.

Lastly, factor the host you're talking to into the concurrency limiter
key. This means that if you're round robining one bad host can't hurt
your collective performance.