[improvement] Provide span operation when using deferred traces (#1014)
## Before this PR
When using versions of tracing-java with https://github.com/palantir/tracing-java/pull/97, the spans created by the concurrency limiter and the dispatcher have the same name as the parent span. This is confusing since the parent span name does not accurately describe the work of the concurrency limiter or the dispatcher.

## After this PR
The concurrency limiter and the dispatcher provide a unique span operation when creating spans.