[improvement] Introduce QoS propagation
Similar to #991, this PR aims to provide a way to propagate 429s and
503s to caller so they can handle the QoS exception.

Unlike #991, the concept of QoS propagation is decoupled from the
backoff and retry configuration to make it more obvious that
IOExceptions will be retried whereas QosExceptions (other than
RetryOther) will be propagated.

The OkHttp layer propagates the QoS by setting the future to the
response and Feign and Retrofit2 will decode the response into the
appropriate QosException.