[fix] Never block threads if a non-IOException is thrown (#1125)
Instead, just wrap in an IOException

We should not be catching errors, but they will just get blackholed.

## Before this PR
If a remote call throws an exception, the Conjure call will block forever. 

## After this PR
Now, fail the call with an exception.


## Possible downsides?
People relying on threads getting permanently blocked for control flow purposes will see a break.