Allow disabling of handling of SocketExceptions
We have an internal service which manages its own multiplexing between a
number of Conjure clients (this was decided as the right approach a
number of months ago). We'd like it to use a client per remote server.

Unfortunately, with Conjure, discovering the server is down requires a 4
second backoff loop.

This PR enables a mode that only we should use which disables retrying
on that, but retains retrying on 503s and 429s. We have a retry loop at
a higher level.