Don't share connection pools
Internally, we've seen a number of issues that appear to be related to connection pool reuse between clients. What we think we see is that a client is hanging onto a connection (which after some period is actually dead, so holding no resources), but the ConnectionPool itself manages a deque of connections, which it synchronizes access too. After long enough runtime, this queue becomes so long that accessing it times out.

Now, the curious thing here is that this connection pool reuse is totally pointless! Reuse relies on the respective connections having the same SSLSocketFactory. And, the way we construct clients internally, we _always_ use a new one each time - because we read the truststore file from disk.