Opt out of object-mapper optimizations by default in jdk16+
Note that this can be overridden on java 16 for experimentation,
however we do not recommend doing so. I considered checking for
`--illegal-access=(permit|warn|debug)` in the jvm args, however
that may be set for other reasons, and we will not test afterburner
on java 16, so it would introduce more risk than it's worth.