Rewrite all tests away from Dropwizard
Dropwizard behaves poorly with GCV since it inadvertently locks
a number of our dependencies into versions that we don't want.
Concretely this has a problem when we are trying to upgrade
our dependencies to the Jakarta EE 9 versions, which only very
recent versions of Dropwizard support, and those versions bring
in dependencies of libraries like slfj4 2.0.0 ALPHA which conflicts
with other services.

This is a large change...I hope I never have to do this again.