Increase string-length observability thresholds
This is meant to increase the signal to noise ratio of the data
we collect in a couple ways:
We collect stack trace information when values are within 20% of
the upcoming limit in Jackson (4mb), helping us focus our efforts
on places that are most likely to be impacted.
We increase the minimum threshold to update our histogram in order
to vastly reduce interactions with the metric because it uses a
recency-biased random sampling reservoir which holds ~1k samples
at a time. With a high rate of interactions, it's likely that we
will fail to sample the largest value's we observe.