Opt out of the TypeFactory cache until we can resolve cache contention
https://github.com/FasterXML/jackson-databind/issues/3876
https://github.com/FasterXML/jackson-benchmarks/pull/5

The linked benchmarks show a small advantage of caching over
not caching in most cases, in the best case for a cache: when
a single operation is repetedly executed. In practice, the cache
is used for all operations and sees heavy evictions, where contention
is more likely and has greater impact on users.

Given the hash collision issues that we're aware of, and the impacts
we've observed, we will remove the cache until a version is available
which isn't impacted by the linked issue for performance that's
more predictable.