Opt out of canonicalization in Conjure JsonFactories
We expect unlimited keyspace due to maps keyed by random IDs which
are usually UUID-based. This interacts poorly with string
canonicalization components in Jackson in ways that cause heavy heap
churn. See the discussion here for the details:
https://github.com/FasterXML/jackson-benchmarks/pull/6

The primary risk is that without canonicalization, we allocate
reused strings much more heavily. Strings used to match setter
methods should be very short lived and should not escape the
parsing thread, so the JIT may be able to do terribly clever
things to help us out.

Note: it may be possible to canonicalize contextually in jackson
3, which would give us the best of both worlds!