SslSocketFactories avoids duplicate effort creating TLS components
Concretely, this simplifies creation of TrustContext and creating
ClientConfiguration from a ServiceConfiguration to load the
TrustManagerFactory from disk a single time rather than once for
the X509TrustManager and a second time for the SSLContext to
create an SSLSocketFactory.

Additionally, we provide public utility functionality such that
consumers may also avoid duplicating efforts with slight code
changes, without reimplementing the logic to create SSLContext
instances themselves.